+++ test/date.js
@@ -9,6 +9,16 @@ describe('Date', function() {
 		});
 	});
 
+	describe('.isDate(variable)', function() {
+		it('should return true if the argument is a date object', function() {
+
+			var date = new Date(),
+			    str  = '';
+			assert.equal(true, Date.isDate(date));
+			assert.equal(false, Date.isDate(str));
+		});
+	});
+
 	describe('#toSource()', function() {
 		it('should return the source code representation of the date', function() {
 			var d = new Date(1);